Kyle: LucasMecanum turning can only work independently from main movement
From Lucas's turning idea, i made it go upon all motors and it turns correctly. But can't turn while moving left stick. Left stick and right stick moving simultaneously won't work. Lucas and I might invest the time to get the simultaneous turning and movement to work. But the question is, do we really need to be able to turn and move at the same time???? Do we need to be that quick for the tournament??
-Kyle and Lucas's Mecanum Code